extracting gesture logic into own function
