+++ app/src/screens/Snake/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from "react";
+import React, { useState } from "react";
 import { Platform, StyleSheet, View } from "react-native";
 
 import ScreenWrapper from "../../components/ScreenWrapper";
@@ -8,15 +8,10 @@ import nextGameState, {
   CardinalDirections,
   GameStatus,
   GameState,
-  Point,
-  NORTH,
-  EAST,
-  SOUTH,
-  WEST
+  EAST
 } from "../../lib/snake";
 import { defaultCellSize } from "../../constants/snake";
 import { getBoardColumnsAndRows, getBoardDimensions } from "../../utils/snake";
-import { getSwipeDirection } from "../../utils/gestures";
 
 import Board from "../../components/SnakeBoard";
 import BoardCell from "../../components/SnakeBoard/Cell";
@@ -25,6 +20,7 @@ import Apple from "../../components/SnakeApple";
 
 import {
   useAnimationFrame,
+  useGestureResponder,
   useNativeKeyboardListener,
   useWebKeyboardListener
 } from "./hooks";
@@ -48,13 +44,6 @@ const defaultState: GameState = {
   columns,
   rows,
   cellCreator: BoardCell
-};
-
-const cardinalDirectionsDictionary = {
-  up: NORTH,
-  right: EAST,
-  down: SOUTH,
-  left: WEST
 };
 
 const SnakeScreen = () => {
@@ -100,24 +89,9 @@ const SnakeScreen = () => {
   }
   /* eslint-enable react-hooks/rules-of-hooks */
 
-  const initialPanPosition = useRef<Point>({ x: 0, y: 0 });
-
-  const onStartShouldSetResponder = ({ nativeEvent }: any) => {
-    initialPanPosition.current = {
-      x: nativeEvent.locationX,
-      y: nativeEvent.locationY
-    };
-    return true;
-  };
-
-  const onResponderRelease = ({ nativeEvent }: any) => {
-    const { x, y } = initialPanPosition.current;
-    const { locationX, locationY } = nativeEvent;
-
-    const swipeDirection = getSwipeDirection(x, y, locationX, locationY);
-    swipeDirection &&
-      onMoveUpdate(cardinalDirectionsDictionary[swipeDirection]);
-  };
+  const { onStartShouldSetResponder, onResponderRelease } = useGestureResponder(
+    onMoveUpdate
+  );
 
   const board = initBoard(defaultState);
   const { width, height } = getBoardDimensions();
