+++ app/web/webpack.config.js
@@ -85,6 +85,7 @@ module.exports = {
   },
 
   devServer: {
+    host: "0.0.0.0",
     port: 3000
   }
 };
